updating & testing ALL existing demos/examples with syntax changes & comments, adding new demos, minor updates ToxiclibsSupport and MeshVoxelizer classes, added Triangle2D.fromBarycentric(), adding more Terrain.toMesh() versions to be more compatible with other primitives
--HG--
branch : toxiclibs

+++ src.p5/toxi/processing/ToxiclibsSupport.java
@@ -325,7 +325,7 @@ public class ToxiclibsSupport {
         }
     }
 
-    public void origin(int len) {
+    public void origin(float len) {
         origin(Vec3D.ZERO, len);
     }
